Bug 27265: fix output_dir value in sub-projects
We update rbm for bug 27265: fix output_dir value in sub-projects.

As part of this fix, the patch in rbm is changing how the value of
output_dir is overriden by rbm when handling input_files: the output_dir
option is now defined under norec, which has priority over normal
options but is not inherided in sub-projects. In rbm.conf we are also
changing the output_dir option in var/build_id_txt so that the build ids
do not depend on the path on the current system. Because of the change
in rbm we now need to reset `norec` so that our value of `output_dir` is
correctly taken into account.