Bug 25099: Set nightly version to tbb-nightly.$year.$month.$day
Include the current day in the nightly version number, so that each
build gets a different version number.

We get this version number using a perl function, and a "state"
variable, so that the day when the build was started is used.
This avoids issues when a build is started on a day but finishs on a
different day.

In the nightly directory, the builds are now stored using the version
instead of just the date.

We also allow overriding this version number by using the
TORBROWSER_NIGHTLY_VERSION environment variable. This could be useful
for example if you need to generate incremental mars for the version
from the previous day.