Bug 25101: Add the prune-old-builds script
Previously we were using the prune-old-builds script which was located
in the testsuite repository in tools/prune-old-builds, as a git submodule.

As this script is not used anywhere other than here, it doesn't make
much sense to keep it as a separate git submodule. To simplify
modification of this script, we move it here.