Bug 33760: Update rbm.conf to match NDK 20
Based on a patch by sisbell, thanks!

We make it clearer that we are concerned with the cross compilation
prefix (so that all the necessary tools can be found) by renaming
`configure_host` to `cross_prefix`.

Additionally, we reorder the `configure_opt` arguments a bit to closer
match other platforms.

Finally, we introduce the Fenix related targets to be able to build
Tor Browser based on Fennec and Fenix in parallel for the time being.