Bug 40015: Use only 64bit fxc2
We use the somewhat surprising breakage due to
https://bugzilla.mozilla.org/show_bug.cgi?id=1619504 to clean up our
fxc2 build and usage.

Previously, we built both the 64bit and 32bit version of fxc2 and were
running them in a 64bit and 32bit Wine environment, respectively.
However, that's not needed because a) Wine is capable of running both
64bit and 32bit binaries and b) fxc2 is only used at build time to
generate shader files. As we are building both Windows flavors in a
64bit container we should easily be able to live with just a 64bit
variant.

We need, nevertheless, take care that we still bundle the correct
d3dcompiler_47.dll library for the respective architecture.