Downgrade NSS and SQLCipher for 61.0.10
We are about to switch to a previous `application-services` version
which is the proper dependency for the `android-components` and `fenix`
commits we'll use. However, that version requires an NSS and SQLCipher
downgrade. The code that landed previously was based on a later
`application-services` commit in the assumption that we'll use a tag
which includes those changes by the time we would wire all our
requirements for `fenix` together.