Bug 40029: Clean-up for mobile
We remove the Fenix targets that were due to Fennec-based builds still
being around. Additionally, we move the *fenix* projects to their
non-Fenix counterparts in case they only existed due to Fennec-based
builds. Thirdly, we do a firefox project clean-up as all the mobile bits
moved to the geckoview project.