Bug 40002: Fix Linux ASan builds for ESR 78
We make the ASan build a first-class tor-browser-build citizen as well.
Additionally, we clean the .mozconfig file up so that we have a ASan
build without the `--enable-debug` and `--enable-test` options. This
makes it more usable for hunting memory corruption bugs. We can think
about separate debug builds that have all the assertions etc. enabled if
we find that part useful.