Bug 40068: Bump versions for 81.1.0b1
We make small adjustments while bumping the versions:

a) We remove the explicit nightly target for `android-components`. If we
need it at some point we can add it back. Meanwhile duplicating the
commits/versions has just additional costs for us when bumping them and
it is not clear anyway that we want to use a non-release commit for
`android-components` in our builds.

b) We adapt the targets in Fenix as all of them got simplified into
`debug`, `nightly`, `beta`, and `release`.