Bug 40101: Pick up 81.1.1 updates
We bump Fenix to 81.1.1 and android-components to 57.0.6 while GeckoView
is staying on 81.0b9 to save us another round of rebasing.

It turns out that declaring a `beta_version` and including that in the
version number for a branch name might not have been the best idea given
that not all our branches are based on mozilla-beta. So, we drop that
and add the beta version directly when needed.